pyproject.toml
[project]
name = "workout-app"
version = "0.1.0"
description = "Generate realistic exercise GIFs using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "pre-commit>=4.5.1",
    "replicate>=1.0.7",
    "typer>=0.24.1",
]

[project.scripts]
generate-exercise-gif = "workout_app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
extend-include = ["*.ipynb"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "notebooks",
]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
unfixable = ["F401"]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
